Updates to Symfony Best Practices
- Moved services.xml to YML
- FrontendBundle and BackendBundle to AppBundle
- Fixed all namespaces
