+++ app/config/config.yml
@@ -67,4 +67,4 @@ swiftmailer:
 fos_user:
     db_driver: orm
     firewall_name: main
-    user_class: Flyers\BackendBundle\Entity\User
+    user_class: AppBundle\Entity\User
