+++ app/config/routing.yml
@@ -1,10 +1,5 @@
-frontend:
-    resource: "@FrontendBundle/Controller/"
-    type:     annotation
-    prefix:   /
-
-backend:
-    resource: "@BackendBundle/Controller/"
+app:
+    resource: "@AppBundle/Controller/"
     type:     annotation
     prefix:   /
