+++ src/AppBundle/AppBundle.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace AppBundle;
+
+use AppBundle\DependencyInjection\Security\Factory\WsseFactory;
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+class AppBundle extends Bundle
+{
+	public function build(ContainerBuilder $container)
+    {
+        parent::build($container);
+
+        $extension = $container->getExtension('security');
+        $extension->addSecurityListenerFactory(new WsseFactory());
+    }
+}
