+++ src/AppBundle/Resources/public/js/filters.js
