+++ src/AppBundle/Resources/public/lib/angular/angular-cookies.js
