+++ src/AppBundle/Resources/public/lib/angular/angular-loader.js
