+++ src/AppBundle/Resources/public/lib/angular/angular-resource.js
