+++ src/AppBundle/Resources/public/lib/angular/version.txt
