--- src/Flyers/BackendBundle/BackendBundle.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Flyers\BackendBundle;
-
-use Flyers\BackendBundle\DependencyInjection\Security\Factory\WsseFactory;
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-
-class BackendBundle extends Bundle
-{
-	public function build(ContainerBuilder $container)
-    {
-        parent::build($container);
-
-        $extension = $container->getExtension('security');
-        $extension->addSecurityListenerFactory(new WsseFactory());
-    }
-}
