--- src/Flyers/BackendBundle/Entity/User.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace Flyers\BackendBundle\Entity;
-
-use FOS\UserBundle\Entity\User as BaseUser;
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity
- * @ORM\Table(name="fos_user")
- */
-class User extends BaseUser
-{
-    /**
-     * @ORM\Id
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
-
-    public function __construct()
-    {
-        parent::__construct();
-    }
-}
