--- src/Flyers/BackendBundle/Resources/config/services.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" ?>
-
-<container xmlns="http://symfony.com/schema/dic/services"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-
-    <!--
-    <parameters>
-        <parameter key="backend.example.class">Flyers\BackendBundle\Example</parameter>
-    </parameters>
-
-    <services>
-        <service id="backend.example" class="%backend.example.class%">
-            <argument type="service" id="service_id" />
-            <argument>plain_value</argument>
-            <argument>%parameter_name%</argument>
-        </service>
-    </services>
-    -->
-
-    <services>
-        <service id="wsse.security.authentication.provider"
-            class="Flyers\BackendBundle\Security\Authentication\Provider\WsseProvider" public="false">
-            <argument /> <!-- User Provider -->
-            <argument>%kernel.cache_dir%/security/nonces</argument>
-        </service>
-
-        <service id="wsse.security.authentication.listener"
-            class="Flyers\BackendBundle\Security\Firewall\WsseListener" public="false">
-            <argument type="service" id="security.context"/>
-            <argument type="service" id="security.authentication.manager" />
-            <argument type="service" id="logger" />
-            <tag name="monolog.logger" />
-            <tag channel="wsse" />
-        </service>
-    </services>
-</container>
