--- src/Flyers/BackendBundle/Resources/views/Default/index.html.twig
@@ -1,1 +1,0 @@
-Hello {{ name }}!
