--- src/Flyers/BackendBundle/Security/Authentication/Token/WsseUserToken.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace Flyers\BackendBundle\Security\Authentication\Token;
-
-use Symfony\Component\Security\Core\Authentication\Token\AbstractToken;
-
-class WsseUserToken extends AbstractToken
-{
-    public $created;
-    public $digest;
-    public $nonce;
-
-    public function __construct(array $roles = array())
-    {
-        parent::__construct($roles);
-
-        // If the user has roles, consider it authenticated
-        $this->setAuthenticated(count($roles) > 0);
-    }
-
-    public function getCredentials()
-    {
-        return '';
-    }
-}
