--- src/Flyers/FrontendBundle/Controller/DefaultController.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Flyers\FrontendBundle\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
-
-class DefaultController extends Controller
-{
-    /**
-     * @Route("/")
-     * @Template()
-     */
-    public function indexAction()
-    {
-        return array();
-    }
-}
