--- src/Flyers/FrontendBundle/FrontendBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Flyers\FrontendBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class FrontendBundle extends Bundle
-{
-}
